core/vdbe: reconcile aggregate inverse after rebase
Upstream and IVM independently introduced AggInverse. Keep one opcode and runtime dispatcher for both window-frame advancement and materialized-view aggregate retractions, and adopt upstream's aggregate collation and IO completion shapes. Refresh the new oracle crate's lockfile entry to the workspace version inherited from main.\n\nTests: cargo test -p turso_core incremental --lib; 267 IVM sqltests; 29 ntile sqltests; cargo test -p sqltest; cargo test -p differential-fuzzer --lib; targeted turso_core clippy

+++ core/incremental/vdbe_maintenance/delta_cursor.rs
@@ -103,7 +103,7 @@ impl DeltaCursor {
                 if !pending.completion.finished() {
                     let completion = pending.completion.clone();
                     self.pending_read = Some(pending);
-                    return Ok(IOResult::IO(IOCompletions::Single(completion)));
+                    return Ok(IOResult::IO(IOCompletions(completion)));
                 }
                 if let Some(error) = pending.completion.get_error() {
                     return Err(LimboError::CompletionError(error));
@@ -158,7 +158,7 @@ impl DeltaCursor {
                     buffer,
                 });
                 if !completion.finished() {
-                    return Ok(IOResult::IO(IOCompletions::Single(completion)));
+                    return Ok(IOResult::IO(IOCompletions(completion)));
                 }
                 continue;
             }
